requirements.txt
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.9.1
boto3==1.13.17
botocore==1.16.17
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.0
cssselect==1.1.0
distlib==0.3.1
dj-database-url==0.4.2
Django==3.0.5
django-cors-headers==3.2.1
django-hashid-field==3.1.3
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
django-storages==1.9.1
django-taggit==1.2.0
django-taggit-serializer==0.1.7
djangorestframework==3.11.0
docutils==0.15.2
filelock==3.0.12
gunicorn==20.0.4
hashids==1.3.1
hyperlink==20.0.1
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
itemadapter==0.1.0
itemloaders==1.0.2
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
lxml==4.5.2
Markdown==3.2.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
parsel==1.6.0
Pillow==7.0.0
Protego==0.1.16
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
queuelib==1.5.0
requests==2.24.0
s3transfer==0.3.3
Scrapy==2.3.0
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
Twisted==20.3.0
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.31
w3lib==1.22.0
zipp==3.1.0
zope.interface==5.1.0
